refactor(packages): Move bibtex parser in its own support file
